refactors edit comment spec tests out of company show
