+++ spec/features/companies/user_sees_one_company_spec.rb
@@ -15,81 +15,4 @@ describe "User sees company show page" do
     expect(page).to have_content("Existing Contacts")
   end
 
-  scenario "a user sees a contact form" do
-    company = Company.create(name: "ESPN")
-    category = Category.create!(title: "Finance")
-    job = company.jobs.create!(title: "Developer", level_of_interest: 70, city: "Denver", category: category)
-    contact = company.contacts.create(name: "Sally", position: "Hero", email: "hera@gmail.com")
-
-    visit company_path(company)
-
-    expect(page).to have_content("Name:")
-    expect(page).to have_content("Position:")
-    expect(page).to have_content("Email:")
-  end
-  scenario "a user fills out a contact" do
-    company = Company.create(name: "ESPN")
-    category = Category.create!(title: "Finance")
-    job = company.jobs.create!(title: "Developer", level_of_interest: 70, city: "Denver", category: category)
-    contact = company.contacts.create(name: "Sally", position: "Hero", email: "hera@gmail.com")
-
-    visit company_path(company)
-
-    fill_in :contact_name, with: "George"
-    fill_in :contact_position, with: "President"
-    fill_in :contact_email, with: "potus@gmail.com"
-
-    click_on "Save"
-
-    expect(current_path).to eq(company_path(company))
-  end
-  scenario "a user sees contact added to contacts" do
-    company = Company.create(name: "ESPN")
-    category = Category.create!(title: "Finance")
-    company.jobs.create(title: "Developer", level_of_interest: 90, city: "Denver", category: category)
-    contact = company.contacts.create(name: "Sally", position: "Hero", email: "hera@gmail.com")
-
-    visit company_path(company)
-
-    fill_in :contact_name, with: "George"
-    fill_in :contact_position, with: "President"
-    fill_in :contact_email, with: "potus@gmail.com"
-
-    click_on "Save"
-
-    expect(current_path).to eq(company_path(company))
-    expect(page).to have_content("George")
-    expect(page).to have_content("President")
-    expect(page).to have_content("potus@gmail.com")
-  end
-  scenario 'user moves to edit contact page' do
-    company = Company.create(name: "ESPN")
-    category = Category.create!(title: "Finance")
-    company.jobs.create(title: "Developer", level_of_interest: 90, city: "Denver", category: category)
-    contact = company.contacts.create(name: "Sally", position: "Hero", email: "hera@gmail.com")
-
-    visit company_path(company)
-
-    click_on 'Edit'
-
-    expect(current_path).to eq(edit_company_contact_path(company, contact))
-
-  end
-  scenario 'user fills in edit contact page and returns to company contact' do
-    company = Company.create(name: "ESPN")
-    category = Category.create!(title: "Finance")
-    company.jobs.create(title: "Developer", level_of_interest: 90, city: "Denver", category: category)
-    contact = company.contacts.create(name: "Sally", position: "Hero", email: "hera@gmail.com")
-
-    visit edit_company_contact_path(company, contact)
-
-    fill_in :contact_name, with: "George"
-    fill_in :contact_position, with: "President"
-    fill_in :contact_email, with: "potus@gmail.com"
-
-    click_on "Save"
-
-    expect(current_path).to eq(company_path(company))
-  end
-
 end
